[song] implement Song
Song is effectively a wrapper around multiple patterns and contains
machinery for managing timers w.r.t. playback (play/pause/stop). It
also contains a primitive callback mechanism to notify callers of
steps.

This callback includes the instruments that need to be played at
that moment in time. Or in the case of bin/four-on-the-floor.js,
it will simply output on stdout.